+++ trunk/client/NmdcHub.h
@@ -197,7 +197,7 @@ public:
 		send("$OpForceMove $Who:" + aUser->getNick() + "$Where:" + aServer + "$Msg:" + aMsg + "|");
 	}
 
-	int getUserCount() const {
+	size_t getUserCount() const {
 		Lock l(cs);
 		return users.size();
 	}
